fixes EntityAlterCraftingFX jitter and skip
- fixes NullPointerExceptions and ClassCastExceptions occurring when the
alter is destroyed while crafting is in progress
The jitter was caused by random motion that is applied during initiation
of the partial effect from the parent constructor.
The skip was caused by the call to super.onUpdate() which set the
previous x, y, and z after onUpdate had already moved the particle.