a start to implementing #157
The DecayTextureStitchHandler and TextureDecay classes are the main
focus of this commit, the way the icons are used for rendering needs
finishing because as of now the alpha of the icon overlay in GUIs and on
the ground are overridden to 1. All decayable tools will need to
implement IDecayable to handle recognizing icons that need an overlay
created.