closes #157
BetweenLands axes, pickaxes, spades, swords and weedwood bow are
decayable. Decay is a value [0,255] stored as "Decay" in item stack NBT
as an integer tag. There are 6 visible texture variants and 6 tooltip
hints to the stage of decay. Putting decayable items in the Purifier
will cleanse them (retaining NBT). The more an item has decayed the
greater the effectiveness and attack damage are reduced. For completely
decayed items their effectiveness and attack damage are reduced by 50%.
The maximum power of a weedwood bow is reduced by 15%. Only items
currently in a players' inventory have the chance to decay. The
probability of a decayable item decaying at all within one second is
4.88%, if the player is in water it is 13.98%, if the player is
currently using or swinging the item the previously stated probabilities
are multiplied by 1.025, the former probability becoming 5% and the
latter 14.33%. The higher a players' decay level the closer the
probability is to being halved.