infuser more mechanics
If left to boil, every 30 seconds the fluid is reduced by 1 bucket due
to evaporation. If an infusion is already made and left to boil dry a
toxic cloud spawns (piggy for debug). Stirring resets the evaporation
counter to 0.
Adding water will reduce the current temp by 1/3 if already heating and
also resets the evaporation count.