infuser added life crystal automation
life crystal added by right clicking on infuser - can be removed by
sneak right clicking but only if the infuser contains an infusion. if an
infusion is about to evaporate and there is a durable life crystal
installed the infuser will auto stir and reset the evaporation counter.
does not auto stir if non infused liquid is evaporating.